SnapdragonCamera: fix thumbnail display
Thumbnail image is usually larger than needed due to the choice of
down-sampling factor. When the decoded bitmap is square, it's width
and height is not correctly set in CircularDrawable, so only the
left-top part of the cropped image is shown. Set the size and only
decode exactly the region for display.

Change-Id: I5f4af66a08f0dc02d84bda605b74c9e2b782fe63
CRs-Fixed: 881895